moved postSubmission into 2 methods of Submission class
