+++ src/Buttons/submission/submissionSeeMore.ts
@@ -6,6 +6,7 @@ export const button: Button = {
 	buttonId: "submissionSeeMore",
 	execute: async (client: Client, interaction: ButtonInteraction) => {
 		const message: Message = interaction.message as Message;
+		
 		if (message.components.length == 2) await message.edit({ components: [submissionButtonsOpen, submissionButtonsVotes] });
 		else await message.edit({ components: [submissionButtonsOpen] });
